check autUser

+++ routes/advisorProfiles.js
@@ -5,9 +5,17 @@ const authenticate = passport.authorize('jwt', { session: false })
 
 router.get('/advisor/:id', authenticate, (req, res, next) => {
     const id = req.params.id
+    userId = req.account._id
     AdvisorProfile.findById(id)
       .then((advisorProfile) => {
         if (!advisorProfile) { return next() }
+
+        if(advisorProfile.user._id !== userId){
+          const error = new Error('Unauthorized')
+          error.status = 401
+          return next(error)
+        }
+
         res.status = 200
         res.json(advisorProfile)
       })
@@ -15,9 +23,16 @@ router.get('/advisor/:id', authenticate, (req, res, next) => {
   })
   .post('/advisor', authenticate, (req, res, next) => {
     let newAdvisorProfile = req.body
-
+    userId = req.account._id
     AdvisorProfile.create(newAdvisorProfile)
       .then((advisorProfile) => {
+
+        if(advisorProfile.user._id !== userId){
+          const error = new Error('Unauthorized')
+          error.status = 401
+          return next(error)
+        }
+
         res.status = 201
         res.json(advisorProfile)
       })
@@ -26,10 +41,17 @@ router.get('/advisor/:id', authenticate, (req, res, next) => {
   .put('/advisor/:id', authenticate, (req, res, next) => {
     const advisorProfileId = req.params.id
     let updaAdvisorProfile = req.body
-
+    userId = req.account._id
     AdvisorProfile.findOneAndUpdate(advisorProfileId, updaAdvisorProfile)
     .then((advisorProfile) => {
       if (!advisorProfile) { return next() }
+
+      if(advisorProfile.user._id !== userId){
+        const error = new Error('Unauthorized')
+        error.status = 401
+        return next(error)
+      }
+
       res.status = 200
       res.json(advisorProfile)
     })
@@ -38,10 +60,17 @@ router.get('/advisor/:id', authenticate, (req, res, next) => {
   .patch('/advisor/:id', authenticate, (req, res, next) => {
     const advisorProfileId = req.params.id
     let updaAdvisorProfile = req.body
-
+    userId = req.account._id
     AdvisorProfile.findOneAndUpdate(advisorProfileId, updaAdvisorProfile)
     .then((advisorProfile) => {
       if (!advisorProfile) { return next() }
+
+      if(advisorProfile.user._id !== userId){
+        const error = new Error('Unauthorized')
+        error.status = 401
+        return next(error)
+      }
+
       res.status = 200
       res.json(advisorProfile)
     })
@@ -50,12 +79,25 @@ router.get('/advisor/:id', authenticate, (req, res, next) => {
   })
   .delete('/advisor/:id', authenticate, (req, res, next) => {
     const id = req.params.id
-    console.log(id)
-    AdvisorProfile.findByIdAndRemove(id)
-    .then((advisorProfile) => {
-      if (!advisorProfile) { return next() }
-      res.status = 204
-    })
+    userId = req.account._id
+
+    AdvisorProfile.findById(id)
+      .then((advisorProfile) => {
+        if (!advisorProfile) { return next() }
+
+        if(advisorProfile.user._id !== userId){
+          const error = new Error('Unauthorized')
+          error.status = 401
+          return next(error)
+        }
+
+        AdvisorProfile.findByIdAndRemove(id)
+        .then((advisorProfile) => {
+          if (!advisorProfile) { return next() }
+          res.status = 204
+        })
+        .catch((error) => next(error))
+      })
     .catch((error) => next(error))
   })
